Change some things around so you can write to the Comment panel in ThreeUp mode without switching to Comment state.
TECHNICALLY this breaks everything related to switching
English/Japanese/Comment states, HOWEVER if you can write directly to
the other panels, this is not necessary for the vast majority of cases.
If I now switch ThreeUp mode to default, remove the *Up toggle, and
change the English/Japanese/Comment buttons to toggle the visibility of
these panels instead, we have not only restored all functionality, but
also made it much easier for the user to a) write comments and b) toggle
the visibility of panels.